[#22] refactor: nanaContent conteont 반환 오류 수정, controller 폴더 생성
